# api/api.yaml
openapi: 3.0.2
servers:
  - url: /api
security:
  - {}
info:
  version: 3.0.0
  title: GripMock API Schema
  description: Documentation and API specification GripMock.
  contact:
    name: Maksim Babichev
    url: https://github.com/radiantbald/gripmock-openapi
  license:
    name: MIT
    url: https://github.com/radiantbald/gripmock-openapi/blob/master/LICENSE
...
paths:
...
  /rooms:
    get:
      tags:
        - rooms
      summary: Room options
      description: Returns distinct non-empty room IDs available in current stubs.
      operationId: roomsList
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rooms'
        '500':
          description: Internal Server Error

  # services reflection
...
  /stubs/{uuid}:
    get:
      tags:
        - stubs
      summary: Get Stub by ID
      description: Searches for Stub by ID
      operationId: findByID
      parameters:
        - name: uuid
          in: path
          description: ID of stub
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stub'
        '400':
          description: Invalid UUID format
        '404':
          description: Stub not found
        '500':
          description: Internal Server Error
    delete:
      tags:
        - stubs
      summary: Deletes stub by ID
      description: The method removes the stub by ID
      operationId: deleteStubByID
      parameters:
        - name: uuid
          in: path
          description: ID of stub
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '204':
          description: successful operation
        '400':
          description: Invalid UUID format
        '404':
          description: Stub not found
        '500':
          description: Internal Server Error
...
components:
  schemas:
...
    ID:
      type: integer
      format: uint64
      minimum: 1
      example: 101
      x-omitzero: false
...
    Rooms:
      type: object
      required:
        - rooms
      properties:
        rooms:
          type: array
          items:
            $ref: '#/components/schemas/Room'
    Room:
      type: object
      required:
        - id
      properties:
        id:
          type: string
        name:
          type: string
...
    Stub:
      type: object
      required:
        - service
        - method
        - input
        - output
      properties:
        id:
          $ref: '#/components/schemas/ID'
        service:
          type: string
          example: Gripmock
          x-omitzero: false
        method:
          type: string
          example: SayHello
          x-omitzero: false
        name:
          type: string
          description: Optional user-defined stub name
        enabled:
          type: boolean
          default: true
          description: Effective enabled status for the selected room context.
          x-go-type-skip-optional-pointer: true
        headers:
          $ref: '#/components/schemas/StubHeaders'
        input:
          $ref: '#/components/schemas/StubInput'
          x-omitzero: false
        inputs:
          type: array
          description: Inputs to match against. If multiple inputs are provided, the stub will be matched if any of the inputs match.
          items:
            $ref: '#/components/schemas/StubInput'
          x-go-type-skip-optional-pointer: true
        output:
          $ref: '#/components/schemas/StubOutput'
          x-omitzero: false
        options:
          $ref: '#/components/schemas/StubOptions'
          x-omitzero: true
        effects:
          type: array
          description: Side effects applied after successful stub match
          items:
            $ref: '#/components/schemas/StubEffect'
          x-go-type-skip-optional-pointer: true
        source:
          type: string
          description: Source of the stub (file, rest, mcp, proxy)
          readOnly: true
          x-omitzero: true
    StubOptions:
      type: object
      description: Optional behavior settings for a stub
      properties:
        times:
          type: integer
          description: Max number of matches; 0 = unlimited
          minimum: 0
          default: 0
          x-go-type-skip-optional-pointer: true
    StubEffect:
      type: object
      required:
        - action
      properties:
        action:
          type: string
          enum: [upsert, delete]
        id:
          type: string
          description: Stub UUID for delete action
          x-go-type-skip-optional-pointer: true
        stub:
          type: object
          additionalProperties: true
          description: Stub payload for upsert action
          x-go-type-skip-optional-pointer: true
    StubInput:
      type: object
      properties:
        ignoreArrayOrder:
          type: boolean
          default: false
          x-go-type-skip-optional-pointer: true
        equals:
          type: object
          additionalProperties: true
          x-go-type-skip-optional-pointer: true
        contains:
          type: object
          additionalProperties: true
          x-go-type-skip-optional-pointer: true
        matches:
          type: object
          additionalProperties: true
          x-go-type-skip-optional-pointer: true
        anyOf:
          type: array
          description: Alternative input matchers (OR logic)
          items:
            $ref: '#/components/schemas/StubInputAnyOfElement'
          x-go-type-skip-optional-pointer: true
    StubInputAnyOfElement:
      type: object
      properties:
        ignoreArrayOrder:
          type: boolean
          default: false
          x-go-type-skip-optional-pointer: true
        equals:
          type: object
          additionalProperties: true
          x-go-type-skip-optional-pointer: true
        contains:
          type: object
          additionalProperties: true
          x-go-type-skip-optional-pointer: true
        matches:
          type: object
          additionalProperties: true
          x-go-type-skip-optional-pointer: true
    StubHeaders:
      type: object
      x-go-type-skip-optional-pointer: true
      properties:
        equals:
          type: object
          additionalProperties:
            type: string
          x-go-type-skip-optional-pointer: true
        contains:
          type: object
          additionalProperties:
            type: string
          x-go-type-skip-optional-pointer: true
        matches:
          type: object
          additionalProperties:
            type: string
          x-go-type-skip-optional-pointer: true
        anyOf:
          type: array
          description: Alternative header matchers (OR logic)
          items:
            $ref: '#/components/schemas/StubHeadersAnyOfElement'
          x-go-type-skip-optional-pointer: true
    StubHeadersAnyOfElement:
      type: object
      properties:
        equals:
          type: object
          additionalProperties:
            type: string
          x-go-type-skip-optional-pointer: true
        contains:
          type: object
          additionalProperties:
            type: string
          x-go-type-skip-optional-pointer: true
        matches:
          type: object
          additionalProperties:
            type: string
          x-go-type-skip-optional-pointer: true
    StubOutput:
      type: object
      properties:
        data:
          type: object
          additionalProperties: true
          x-go-type-skip-optional-pointer: true
        stream:
          type: array
          items:
            type: object
            additionalProperties: true
          x-go-type-skip-optional-pointer: true
        headers:
          type: object
          additionalProperties:
            type: string
          x-go-type-skip-optional-pointer: true
        error:
          type: string
          example: Message not found
          x-go-type-skip-optional-pointer: true
        code:
          type: integer
          format: uint32
          x-go-type: codes.Code
          x-go-type-import:
            name: codes
            path: google.golang.org/grpc/codes
          example: 3
          x-go-type-skip-optional-pointer: true
        details:
          type: array
          description: gRPC status details packed into google.protobuf.Any (each item must contain type URL in `type`)
          items:
            type: object
            required: [type]
            properties:
              type:
                type: string
                description: Full Any type URL (for example, type.googleapis.com/google.rpc.ErrorInfo)
            additionalProperties: true
          x-go-type-skip-optional-pointer: true
        delay:
          type: string
          x-go-type: gptypes.Duration
          x-go-type-import:
            name: gptypes
            path: github.com/radiantbald/gripmock/v3/internal/infra/types
          description: Delay before sending the response
          example: "1s"
          x-omitzero: true
          x-go-type-skip-optional-pointer: true